ResultData에 Data1Name 추가

+++ src/main/java/com/KoreaIT/smw/demo/controller/UsrArticleController.java
@@ -23,28 +23,61 @@ public class UsrArticleController {
 	// 액션메서드
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData<Integer> doModify(int id, String title, String body) {
-		Article article = articleService.getArticle(id);
-		if (article == null) {
-			return ResultData.from("F-1", Ut.f("%d번 글은 존재하지 않습니다", id), id);
+	public ResultData<Integer> doModify(HttpSession httpSession, int id, String title, String body) {
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
 		}
 
-		articleService.modifyArticle(id, title, body);
+		if (isLogined == false) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
+		}
 
-		return ResultData.from("S-1", Ut.f("%d번 글을 수정 했습니다", id), id);
+		Article article = articleService.getArticle(id);
+		if (article == null) {
+			return ResultData.from("F-1", Ut.f("%d번 글은 존재하지 않습니다", id), "id", id);
+		}
+
+		ResultData actorCanModifyRd = articleService.actorCanModify(loginedMemberId, article);
+
+		if (actorCanModifyRd.isFail()) {
+			return actorCanModifyRd;
+		}
+
+		return articleService.modifyArticle(id, title, body);
+
 	}
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData<Integer> doDelete(int id) {
+	public ResultData<Integer> doDelete(HttpSession httpSession, int id) {
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+
+		if (isLogined == false) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
+		}
+
 		Article article = articleService.getArticle(id);
 		if (article == null) {
-			return ResultData.from("F-1", Ut.f("%d번 글은 존재하지 않습니다", id), id);
+			return ResultData.from("F-1", Ut.f("%d번 글은 존재하지 않습니다", id), "id", id);
+		}
+
+		if (article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-2", Ut.f("%d번 글에 대한 권한이 없습니다", id));
 		}
 
 		articleService.deleteArticle(id);
 
-		return ResultData.from("S-1", Ut.f("%d번 글을 삭제 했습니다", id), id);
+		return ResultData.from("S-1", Ut.f("%d번 글을 삭제 했습니다", id), "id", id);
 	}
 
 	@RequestMapping("/usr/article/doWrite")
@@ -76,7 +109,7 @@ public class UsrArticleController {
 
 		Article article = articleService.getArticle(id);
 
-		return ResultData.newData(writeArticleRd, article);
+		return ResultData.newData(writeArticleRd, "article", article);
 	}
 
 	@RequestMapping("/usr/article/getArticles")
@@ -84,7 +117,7 @@ public class UsrArticleController {
 	public ResultData<List<Article>> getArticles() {
 		List<Article> articles = articleService.articles();
 
-		return ResultData.from("S-1", "Article List", articles);
+		return ResultData.from("S-1", "Article List", "List<Article>", articles);
 	}
 
 	@RequestMapping("/usr/article/getArticle")
@@ -97,7 +130,7 @@ public class UsrArticleController {
 			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다", id));
 		}
 
-		return ResultData.from("S-1", Ut.f("%d번 게시물입니다", id), article);
+		return ResultData.from("S-1", Ut.f("%d번 게시물입니다", id), "article", article);
 	}
 
 }
